move CreaetSource function to ProtocolUtils instead of JSONUtils
